Refactored utility functions and cleaned up the life cycle test module.
