+++ RUFAS/simulation_engine.py
@@ -5,7 +5,7 @@ import time as timer
 from pathlib import Path
 from RUFAS import routines, errors, classes
 from RUFAS.classes import Config, State, Weather, Time
-from RUFAS.util import get_base_dir, read_json_file
+from RUFAS.util import Utility
 from RUFAS.output_handler import OutputHandler
 import random
 import numpy
@@ -141,7 +141,7 @@ class SimulationEngine:
 
         print(f"Initializing simulation environment from {file_path}")
         try:
-            data = read_json_file(file_path)
+            data = Utility.read_json_file(file_path)
             self.config = Config(data['config'], data['weather'])
 
             if self.config.set_seed:
@@ -152,8 +152,8 @@ class SimulationEngine:
             self.time = Time(self.config)
             self.state = State(data['farm'], self.config,
                                self.weather, self.time)
-            self.output = OutputHandler(classes.read_json_file(
-                get_base_dir() / 'input/output' / data['output']), self.state)
+            self.output = OutputHandler(Utility.read_json_file(
+                Utility.get_base_dir() / 'input/output' / data['output']), self.state)
 
         except errors.JSONfileData as e:
             print(
