+++ tests/test_animal/test_life_cycle.py
@@ -12,9 +12,9 @@ from RUFAS.routines.animal.life_cycle.cow import Cow
 from RUFAS.routines.animal.life_cycle.heiferI import HeiferI
 from RUFAS.routines.animal.life_cycle.heiferII import HeiferII
 from RUFAS.routines.animal.life_cycle.heiferIII import HeiferIII
-from RUFAS.routines.animal.life_cycle.life_cycle import calc_average, \
-    LifeCycleManager, percent_calculator, remove_items_from_list_by_indices
+from RUFAS.routines.animal.life_cycle.life_cycle import LifeCycleManager
 from RUFAS.routines.animal.pen import Pen
+from RUFAS.util import Utility
 
 
 @fixture
@@ -38,8 +38,9 @@ def life_cycle_manager(mocker: MockerFixture) -> LifeCycleManager:
     return lcm
 
 
-def test__set_avg_CI_should_set_custom_avg_CI_given_user_input_CI(mocker: MockerFixture,
-                                                                  life_cycle_manager) -> None:
+def test_set_avg_CI(mocker: MockerFixture, life_cycle_manager) -> None:
+    # Given user input calving interval,
+    # the function should correctly set the custom average calving interval.
     mock_animal_config_data = {
         'user_input_calving_interval': True,
         'calving_interval': 400.0
@@ -48,17 +49,18 @@ def test__set_avg_CI_should_set_custom_avg_CI_given_user_input_CI(mocker: Mocker
     mock_animal_config.__getitem__.side_effect = mock_animal_config_data.__getitem__
     mock_animal_initializer = mocker.MagicMock(autospec=AnimalInitialization)
 
-    spy__set_avg_CI = mocker.spy(life_cycle_manager, '_set_avg_CI')
+    spy_set_avg_CI = mocker.spy(life_cycle_manager, '_set_avg_CI')
     assert life_cycle_manager.avg_CI == approx(0.0)
 
     life_cycle_manager._set_avg_CI(mock_animal_config, mock_animal_initializer)
 
     assert life_cycle_manager.avg_CI == approx(400.0)
-    spy__set_avg_CI.assert_called_once()
+    spy_set_avg_CI.assert_called_once()
 
 
-def test__set_avg_CI_should_use_cow_avg_CI_given_no_user_input_CI(mocker: MockerFixture,
-                                                                  life_cycle_manager) -> None:
+def test_set_avg_CI_2(mocker: MockerFixture, life_cycle_manager) -> None:
+    # Given no user input calving interval,
+    # the function should use the cow average calving interval.
     mock_animal_config_data = {
         'user_input_calving_interval': False
     }
@@ -71,17 +73,17 @@ def test__set_avg_CI_should_use_cow_avg_CI_given_no_user_input_CI(mocker: Mocker
     mock_animal_initializer = mocker.MagicMock(autospec=AnimalInitialization)
     mock_animal_initializer.initialization_db_summary.return_value = mock_init_db_summary
 
-    spy__get_avg_CI = mocker.spy(life_cycle_manager, '_set_avg_CI')
+    spy_get_avg_CI = mocker.spy(life_cycle_manager, '_set_avg_CI')
     assert life_cycle_manager.avg_CI == approx(0.0)
 
     life_cycle_manager._set_avg_CI(mock_animal_config, mock_animal_initializer)
 
     assert life_cycle_manager.avg_CI == approx(500.0)
-    spy__get_avg_CI.assert_called_once()
+    spy_get_avg_CI.assert_called_once()
     mock_animal_initializer.initialization_db_summary.assert_called_once()
 
 
-def test__get_calves(mocker: MockerFixture, life_cycle_manager) -> None:
+def test_get_calves(mocker: MockerFixture, life_cycle_manager) -> None:
     calf_num = 10
     breed = 'HO'
     days_born = 50
@@ -93,7 +95,7 @@ def test__get_calves(mocker: MockerFixture, life_cycle_manager) -> None:
         mock_calves.append(mock_new_calf)
     mock_animal_initializer = mocker.MagicMock(autospec=AnimalInitialization)
     mock_animal_initializer.get_calves.return_value = mock_calves
-    spy__get_calves = mocker.spy(life_cycle_manager, '_get_calves')
+    spy_get_calves = mocker.spy(life_cycle_manager, '_get_calves')
 
     calves = life_cycle_manager._get_calves(calf_num, breed, mock_animal_initializer)
 
@@ -102,11 +104,11 @@ def test__get_calves(mocker: MockerFixture, life_cycle_manager) -> None:
         assert calf.breed == breed
         calf.events.add_event.assert_called_once_with(days_born, 0, INIT_HERD)
 
-    spy__get_calves.assert_called_once_with(calf_num, breed, mock_animal_initializer)
+    spy_get_calves.assert_called_once_with(calf_num, breed, mock_animal_initializer)
     mock_animal_initializer.get_calves.assert_called_once_with(calf_num, breed)
 
 
-def test__get_heiferIs(mocker: MockerFixture, life_cycle_manager) -> None:
+def test_get_heiferIs(mocker: MockerFixture, life_cycle_manager) -> None:
     heiferI_num = 10
     breed = 'HO'
     days_born = 350
@@ -118,7 +120,7 @@ def test__get_heiferIs(mocker: MockerFixture, life_cycle_manager) -> None:
         mock_heiferIs.append(mock_new_heiferI)
     mock_animal_initializer = mocker.MagicMock(autospec=AnimalInitialization)
     mock_animal_initializer.get_heiferIs.return_value = mock_heiferIs
-    spy__get_heiferIs = mocker.spy(life_cycle_manager, '_get_heiferIs')
+    spy_get_heiferIs = mocker.spy(life_cycle_manager, '_get_heiferIs')
 
     heiferIs = life_cycle_manager._get_heiferIs(heiferI_num, breed, mock_animal_initializer)
 
@@ -127,11 +129,11 @@ def test__get_heiferIs(mocker: MockerFixture, life_cycle_manager) -> None:
         assert heiferI.breed == breed
         heiferI.events.add_event.assert_called_once_with(days_born, 0, INIT_HERD)
 
-    spy__get_heiferIs.assert_called_once_with(heiferI_num, breed, mock_animal_initializer)
+    spy_get_heiferIs.assert_called_once_with(heiferI_num, breed, mock_animal_initializer)
     mock_animal_initializer.get_heiferIs.assert_called_once_with(heiferI_num, breed)
 
 
-def test__get_heiferIIs(mocker: MockerFixture, life_cycle_manager) -> None:
+def test_get_heiferIIs(mocker: MockerFixture, life_cycle_manager) -> None:
     heiferII_num = 10
     breed = 'HO'
     days_born = 800
@@ -143,7 +145,7 @@ def test__get_heiferIIs(mocker: MockerFixture, life_cycle_manager) -> None:
         mock_heiferIIs.append(mock_new_heiferIII)
     mock_animal_initializer = mocker.MagicMock(autospec=AnimalInitialization)
     mock_animal_initializer.get_heiferIIs.return_value = mock_heiferIIs
-    spy__get_heiferIIs = mocker.spy(life_cycle_manager, '_get_heiferIIs')
+    spy_get_heiferIIs = mocker.spy(life_cycle_manager, '_get_heiferIIs')
 
     heiferIIs = life_cycle_manager._get_heiferIIs(heiferII_num, breed, mock_animal_initializer)
 
@@ -152,11 +154,11 @@ def test__get_heiferIIs(mocker: MockerFixture, life_cycle_manager) -> None:
         assert heiferII.breed == breed
         heiferII.events.add_event.assert_called_once_with(days_born, 0, INIT_HERD)
 
-    spy__get_heiferIIs.assert_called_once_with(heiferII_num, breed, mock_animal_initializer)
+    spy_get_heiferIIs.assert_called_once_with(heiferII_num, breed, mock_animal_initializer)
     mock_animal_initializer.get_heiferIIs.assert_called_once_with(heiferII_num, breed)
 
 
-def test__get_heiferIIIs(mocker: MockerFixture, life_cycle_manager) -> None:
+def test_get_heiferIIIs(mocker: MockerFixture, life_cycle_manager) -> None:
     heiferIII_num = 10
     breed = 'HO'
     days_born = 650
@@ -168,7 +170,7 @@ def test__get_heiferIIIs(mocker: MockerFixture, life_cycle_manager) -> None:
         mock_heiferIIIs.append(mock_new_heiferIII)
     mock_animal_initializer = mocker.MagicMock(autospec=AnimalInitialization)
     mock_animal_initializer.get_heiferIIIs.return_value = mock_heiferIIIs
-    spy__get_heiferIIIs = mocker.spy(life_cycle_manager, '_get_heiferIIIs')
+    spy_get_heiferIIIs = mocker.spy(life_cycle_manager, '_get_heiferIIIs')
 
     heiferIIIs = life_cycle_manager._get_heiferIIIs(heiferIII_num, breed, mock_animal_initializer)
 
@@ -177,11 +179,11 @@ def test__get_heiferIIIs(mocker: MockerFixture, life_cycle_manager) -> None:
         assert heiferIII.breed == breed
         heiferIII.events.add_event.assert_called_once_with(days_born, 0, INIT_HERD)
 
-    spy__get_heiferIIIs.assert_called_once_with(heiferIII_num, breed, mock_animal_initializer)
+    spy_get_heiferIIIs.assert_called_once_with(heiferIII_num, breed, mock_animal_initializer)
     mock_animal_initializer.get_heiferIIIs.assert_called_once_with(heiferIII_num, breed)
 
 
-def test__get_cows(mocker: MockerFixture, life_cycle_manager) -> None:
+def test_get_cows(mocker: MockerFixture, life_cycle_manager) -> None:
     cow_num = 10
     breed = 'HO'
     days_born = 2200
@@ -193,7 +195,7 @@ def test__get_cows(mocker: MockerFixture, life_cycle_manager) -> None:
         mock_cows.append(mock_new_cow)
     mock_animal_initializer = mocker.MagicMock(autospec=AnimalInitialization)
     mock_animal_initializer.get_cows.return_value = mock_cows
-    spy__get_cows = mocker.spy(life_cycle_manager, '_get_cows')
+    spy_get_cows = mocker.spy(life_cycle_manager, '_get_cows')
 
     cows = life_cycle_manager._get_cows(cow_num, breed, mock_animal_initializer)
 
@@ -202,7 +204,7 @@ def test__get_cows(mocker: MockerFixture, life_cycle_manager) -> None:
         assert cow.breed == breed
         cow.events.add_event.assert_called_once_with(days_born, 0, INIT_HERD)
 
-    spy__get_cows.assert_called_once_with(cow_num, breed, mock_animal_initializer)
+    spy_get_cows.assert_called_once_with(cow_num, breed, mock_animal_initializer)
     mock_animal_initializer.get_cows.assert_called_once_with(cow_num, breed)
 
 
@@ -227,30 +229,30 @@ def test_initialize_herd(mocker: MockerFixture, life_cycle_manager, herd_data: H
     mocker.patch('RUFAS.routines.animal.life_cycle.life_cycle.AnimalInitialization',
                  return_value=mock_animal_initializer)
 
-    mock__set_avg_CI = mocker.patch.object(life_cycle_manager, '_set_avg_CI')
-    mock__get_calves = mocker.patch.object(life_cycle_manager, '_get_calves')
-    mock__get_heiferIs = mocker.patch.object(life_cycle_manager, '_get_heiferIs')
-    mock__get_heiferIIs = mocker.patch.object(life_cycle_manager, '_get_heiferIIs')
-    mock__get_heiferIIIs = mocker.patch.object(life_cycle_manager, '_get_heiferIIIs')
-    mock__get_cows = mocker.patch.object(life_cycle_manager, '_get_cows')
+    mock_set_avg_CI = mocker.patch.object(life_cycle_manager, '_set_avg_CI')
+    mock_get_calves = mocker.patch.object(life_cycle_manager, '_get_calves')
+    mock_get_heiferIs = mocker.patch.object(life_cycle_manager, '_get_heiferIs')
+    mock_get_heiferIIs = mocker.patch.object(life_cycle_manager, '_get_heiferIIs')
+    mock_get_heiferIIIs = mocker.patch.object(life_cycle_manager, '_get_heiferIIIs')
+    mock_get_cows = mocker.patch.object(life_cycle_manager, '_get_cows')
 
     results = life_cycle_manager.initialize_herd(herd_num, calf_num, heiferI_num, heiferII_num,
                                                  heiferIII_num, cow_num, replace_num, herd_init,
                                                  breed, mock_config)
 
     assert life_cycle_manager.herd_num == herd_data['herd_num']
-    mock__set_avg_CI.assert_called_once_with(mock_animal_config, mock_animal_initializer)
-    mock__get_calves.assert_called_once_with(calf_num, breed, mock_animal_initializer)
-    mock__get_heiferIs.assert_called_once_with(heiferI_num, breed, mock_animal_initializer)
-    mock__get_heiferIIs.assert_called_once_with(heiferII_num, breed, mock_animal_initializer)
-    mock__get_heiferIIIs.assert_called_once_with(heiferIII_num, breed, mock_animal_initializer)
-    mock__get_cows.assert_called_once_with(cow_num, breed, mock_animal_initializer)
+    mock_set_avg_CI.assert_called_once_with(mock_animal_config, mock_animal_initializer)
+    mock_get_calves.assert_called_once_with(calf_num, breed, mock_animal_initializer)
+    mock_get_heiferIs.assert_called_once_with(heiferI_num, breed, mock_animal_initializer)
+    mock_get_heiferIIs.assert_called_once_with(heiferII_num, breed, mock_animal_initializer)
+    mock_get_heiferIIIs.assert_called_once_with(heiferIII_num, breed, mock_animal_initializer)
+    mock_get_cows.assert_called_once_with(cow_num, breed, mock_animal_initializer)
     mock_animal_initializer.get_replacement_cows.assert_called_once_with(replace_num, breed)
     assert life_cycle_manager.replacement_market == mock_replacement_cows
     assert len(results) == 5
 
 
-def test__reset_parity(life_cycle_manager) -> None:
+def test_reset_parity(life_cycle_manager) -> None:
     parities = LifeCycleManager.num_cow_for_parity
     preg_times = LifeCycleManager.avg_calving_to_preg_time
     count_per_parity = 10
@@ -276,7 +278,7 @@ def test__reset_parity(life_cycle_manager) -> None:
         assert life_cycle_manager.avg_age_for_calving[parity] == approx(0.0)
 
 
-def test__reset_cull_reason_stats(life_cycle_manager: LifeCycleManager) -> None:
+def test_reset_cull_reason_stats(life_cycle_manager: LifeCycleManager) -> None:
     stats = LifeCycleManager.cull_reason_stats
     num_reasons = len(stats)
     count_per_reason = 10
@@ -293,7 +295,7 @@ def test__reset_cull_reason_stats(life_cycle_manager: LifeCycleManager) -> None:
         assert life_cycle_manager.cull_reason_stats_percent[cull_reason] == approx(0.0)
 
 
-def test__wean_calf(mocker: MockerFixture) -> None:
+def test_wean_calf(mocker: MockerFixture) -> None:
     calf_body_weight_history = [10.0, 20.0]
     calf_pen_history = [mocker.MagicMock(autospec=Pen) for _ in range(3)]
 
@@ -318,7 +320,7 @@ def test__wean_calf(mocker: MockerFixture) -> None:
     assert heiferIs[0] == mock_new_heiferI  # check identity
 
 
-def test__calf_to_heiferI(mocker: MockerFixture, life_cycle_manager: LifeCycleManager) -> None:
+def test_calf_to_heiferI(mocker: MockerFixture, life_cycle_manager: LifeCycleManager) -> None:
     sim_day = 10
     mock_weaned_calf = mocker.MagicMock(autospec=Calf)
     mock_weaned_calf.update.return_value = True
@@ -348,142 +350,103 @@ def test__calf_to_heiferI(mocker: MockerFixture, life_cycle_manager: LifeCycleMa
     assert life_cycle_manager.avg_mature_body_weight == approx(200.0)
 
 
-def test__move_heiferI_to_second_stage(mocker: MockerFixture) -> None:
-    heiferI_body_weight_history = [30.0, 40.0]
-    heiferI_pen_history = [mocker.MagicMock(autospec=Pen) for _ in range(3)]
-    heifer_repro_method = 'TAI'
-    heifer_TAI_protocol = 'd5CG2P'
-    heifer_synchED_protocol = '2P'
-
-    heiferI = mocker.MagicMock(autospec=HeiferI)
-    heiferI.body_weight_history = heiferI_body_weight_history
-    heiferI.pen_history = heiferI_pen_history
-    heiferI.repro_program = heifer_repro_method
-    heiferI.tai_method_h = heifer_TAI_protocol
-    heiferI.synch_ed_method_h = heifer_synchED_protocol
-
-    mock_args = mocker.MagicMock(autospec=Dict)
-    heiferI.get_calf_values.return_value = mock_args
-
-    heiferIIs: List[HeiferII] = []
-    mock_new_heiferII = mocker.MagicMock(autospec=HeiferII)
-    mocker.patch('RUFAS.routines.animal.life_cycle.life_cycle.HeiferII',
-                 return_value=mock_new_heiferII)
-    # LifeCycleManager._move_heiferI_to_second_stage(heiferI, heiferIIs)
-    #
-    # mock_args.update.assert_called_once_with({
-    #     'body_weight_history': heiferI_body_weight_history,
-    #     'pen_history': heiferI_pen_history
-    # })
-    # mock_args.update.assert_called_once_with({
-    #     'repro_program': heifer_repro_method
-    # })
-    # mock_args.update.assert_called_once_with({
-    #     'tai_method_h': heifer_TAI_protocol
-    # })
-    # mock_args.update.assert_called_once_with({
-    #     'synch_ed_method_h': heifer_synchED_protocol
-    # })
-    # assert len(heiferIIs) == 1
-    # assert heiferIIs[0] == mock_new_heiferII  # check identity
-
-
-def test_calc_average_should_pass_given_valid_inputs() -> None:
-    result = calc_average(num_values=9, cur_avg=5, new_value=6)
+def test_calc_average() -> None:
+    # Normal case
+    result = Utility.calc_average(num_values=9, cur_avg=5, new_value=6)
     actual_new_num_values, actual_new_avg = result
     assert actual_new_num_values == 10
     assert actual_new_avg == approx(5.1)  # (9 * 5 + 6) / 10
 
-
-def test_calc_average_should_pass_given_zero_count_and_zero_avg() -> None:
-    result = calc_average(num_values=0, cur_avg=0.0, new_value=6.0)
+    # Given a count of 0 and an average value of 0.0,
+    # the function should return whatever the new value is.
+    result = Utility.calc_average(num_values=0, cur_avg=0.0, new_value=6.0)
     actual_new_num_values, actual_new_avg = result
     assert actual_new_num_values == 1
     assert actual_new_avg == approx(6.0)
 
-
-def test_calc_average_should_pass_given_one_count_and_zero_avg() -> None:
-    result = calc_average(num_values=1, cur_avg=0.0, new_value=6.0)
+    # Given a count of 1 and an average value of 0.0,
+    # the function should pass.
+    result = Utility.calc_average(num_values=1, cur_avg=0.0, new_value=6.0)
     actual_new_num_values, actual_new_avg = result
     assert actual_new_num_values == 2
     assert actual_new_avg == approx(3.0)
 
 
-def test_remove_items_from_list_by_indices_should_do_nothing_given_empty_list_and_empty_removed_indices() -> None:
+def test_remove_items_from_list_by_indices() -> None:
+    # Given an empty list and an empty list of removal indices,
+    # the function should do nothing.
     arr = []
-    removed_idx = []
-    remove_items_from_list_by_indices(arr, removed_idx)
+    del_idx = []
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
     assert len(arr) == 0
 
-
-def test_remove_items_from_list_by_indices_should_do_nothing_given_nonempty_list_and_empty_removed_indices() -> None:
+    # Given a non-empty list and an empty list of removal indices,
+    # the function should do nothing.
     arr = [0, 1, 2]
-    removed_idx = []
-    remove_items_from_list_by_indices(arr, removed_idx)
+    del_idx = []
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
     assert arr == [0, 1, 2]
 
-
-def test_remove_items_from_list_by_indices_should_get_empty_list_given_sized_one_list_and_one_removed_index() -> None:
+    # Given a list of size 1 and the removal index of 0,
+    # the function should return an empty list.
     arr = [0]
-    removed_idx = [0]
-    remove_items_from_list_by_indices(arr, removed_idx)
+    del_idx = [0]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
     assert len(arr) == 0
 
-
-def test_remove_items_from_list_by_indices_should_pass_given_sized_two_list_and_one_removed_index() -> None:
+    # Given a list of size 2 and one valid removal index,
+    # the function should return a correct list of size 1.
     arr = [10, 20]
-    removed_idx = [0]
-    remove_items_from_list_by_indices(arr, removed_idx)
+    del_idx = [0]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
     assert arr == [20]
 
     arr = [10, 20]
-    removed_idx = [1]
-    remove_items_from_list_by_indices(arr, removed_idx)
+    del_idx = [1]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
     assert arr == [10]
 
-
-def test_remove_items_from_list_by_indices_should_pass_given_sized_three_list_and_two_removed_indices() -> None:
+    # Given a list of size 3 and a list of 2 removal indices,
+    # the function should return a correct list of size 1.
     arr = [10, 20, 30]
-    removed_idx = [0, 1]
-    remove_items_from_list_by_indices(arr, removed_idx)
+    del_idx = [0, 1]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
     assert arr == [30]
 
     arr = [10, 20, 30]
-    removed_idx = [1, 2]
-    remove_items_from_list_by_indices(arr, removed_idx)
+    del_idx = [1, 2]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
     assert arr == [10]
 
     arr = [10, 20, 30]
-    removed_idx = [0, 2]
-    remove_items_from_list_by_indices(arr, removed_idx)
+    del_idx = [0, 2]
+    Utility.remove_items_from_list_by_indices(arr, del_idx)
     assert arr == [20]
 
-
-def test_remove_items_from_list_by_indices_should_raise_index_error_given_empty_list_and_nonempty_removed_indices() \
-        -> None:
+    # Given an empty list and a non-empty list of removal indices,
+    # the function should raise IndexError.
     arr = []
-    removed_idx = [0]
+    del_idx = [0]
     with raises(IndexError):
-        remove_items_from_list_by_indices(arr, removed_idx)
+        Utility.remove_items_from_list_by_indices(arr, del_idx)
 
 
-def test_percent_calculator_should_get_numerator_as_pct_given_denominator_100() -> None:
-    pc = percent_calculator(denominator=100)
+def test_percent_calculator() -> None:
+    # Given denominator 100, we should get the numerator as percentage.
+    pc = Utility.percent_calculator(denominator=100)
     assert pc(0.0) == approx(0.0)
     assert pc(12.3) == approx(12.3)
     assert pc(100.0) == approx(100.0)
 
-
-def test_percent_calculator_should_pass_given_valid_denominator() -> None:
-    pc = percent_calculator(denominator=20)
+    # Given any random non-zero denominator, we should get correct percentages.
+    pc = Utility.percent_calculator(denominator=20)
     assert pc(0) == approx(0.0)
     assert pc(20) == approx(100.0)
     assert pc(8) == approx(40.0)
     assert pc(-8) == approx(-40.0)
     assert pc(24) == approx(120.0)
 
-
-def test_percent_calculator_should_raise_zero_division_error_given_denominator_0() -> None:
-    pc = percent_calculator(denominator=0)
+    # Given 0 denominator, we should get a ZeroDivisionError.
+    pc = Utility.percent_calculator(denominator=0)
     with raises(ZeroDivisionError):
         pc(1.0)
